+++ src/ml/data_processing/dispatcher.py
@@ -8,12 +8,8 @@ Date: 06/29/2025
 
 import src.ml.data_processing.signals as sig
 import src.ml.data_processing.technicals as te
-from pandas import DataFrame, Series
+import pandas as pd
 from typing import Dict, Any
-
-
-# delta, diff objects are handled seperately since they 
-# are special cases
 
 FEATURE_DISPATCH = {
     "SMA": te.sma,
@@ -28,9 +24,12 @@ LABEL_DISPATCH = {
     "below": sig.below
 }
 
-def dispatch_feature(dispatch: str, df: DataFrame, json: Dict[str, Any]) -> Series:
+def dispatch_feature(dispatch: str, df: pd.DataFrame, 
+                     json: Dict[str, Any]) -> pd.Series:
+    '''Dispatcher for technical indicators (features)'''
     return FEATURE_DISPATCH[dispatch](df, json, col="Close")
 
-def dispatch_label(dispatch: str, df: DataFrame, col1: str, 
-                   col2: str) -> Series:
+def dispatch_label(dispatch: str, df: pd.DataFrame, col1: str, 
+                   col2: str) -> pd.Series:
+    '''Dispatcher for labeling logic from technical indicators (features)'''
     return LABEL_DISPATCH[dispatch](df, col1, col2)